Initial support of inverse of InterpolatedGeocentricTransform. This inverse operation require an iterative process during the coordinate transformation. For now, we iterate only once as suggested in the NTG_88 document.
git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK8@1715451 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-referencing/src/main/java/org/apache/sis/referencing/operation/transform/AbstractMathTransform.java
@@ -569,7 +569,7 @@ public abstract class AbstractMathTransform extends FormattableObject
                 }
             }
             for (int i=0; i<srcStop; i++) {
-                buffer[bufferedSrcOff + i] = (double) srcPts[srcOff + i];
+                buffer[bufferedSrcOff + i] = srcPts[srcOff + i];
             }
             assert !IterationStrategy.suggest(bufferedSrcOff, dimSource, 0, dimTarget, numBufferedPts).needBuffer;
             try {
@@ -714,7 +714,7 @@ public abstract class AbstractMathTransform extends FormattableObject
                 dstLength = numPts * dimTarget;
             }
             for (int i=0; i<srcLength; i++) {
-                buffer[i] = (double) srcPts[srcOff++];
+                buffer[i] = srcPts[srcOff++];
             }
             try {
                 transform(buffer, 0, dstPts, dstOff, numBufferedPts);
